requirements.txt
backports.ssl-match-hostname==3.5.0.1
decorator==4.0.10
Django==1.10.5
djangorestframework==3.5.1
duo-client==3.0
netaddr==0.7.18
psycopg2==2.6.2
pyldaplite==0.1.3
pysaml2==4.5.0
python-ldap==2.4.28
requests==2.18.4
six==1.11.0
validators==0.11.0
